Improve userscript error handling
Resolve #665.

The current behavior of merely logging a warning when a hook fails is
not ideal. For many (if not all) use cases a failure of these hooks
effectively means a failure of the build because the output files are
not aligned with intent.

An improvement of this behavior should have the following
characteristics:
- facilitate programmatic detection of failures without parsing the
  output
- avoid building images for devices/branches that already have failed
  hooks

This PR tries to strike a balance between exiting immediately on failure
and continuing the build matrix:
- if begin.sh fails, end the process immediately
- if before.sh fails, skip the branch
- if pre-build.sh fails, skip the branch/device
- if any userscript fails, give a non-zero exit code

With this change, I would also question whether running post-build.sh
twice is still necessary for visibility.

I tested the changes by creating all 5 userscript files, setting them to
`exit 0`, running the build and confirming it succeeded. I then re-ran
the build 5 times, each time setting one of the files to `exit 1` and
verifying that the build failed with appropriate output.